[POTATO-8] 메인페이지 링크 연결 및 메뉴 컴포넌트로 분리
